c++ solution of the problem/Programming task
Write a function called swap() that interchanges two int values passed to it by the calling program. (Note that this function swaps the values of the variables in the calling program, i.e., the original. You’ll need to decide how to pass the arguments. Create a main() program to exercise the function.